+++ services/graphics-server/src/blitstr2.rs
@@ -5,10 +5,10 @@ pub use cliprect::*;
 pub(crate) mod fonts;
 pub(crate) use fonts::*;
 
-const LINES: usize = crate::backend::FB_LINES;
-const WIDTH: usize = crate::backend::FB_WIDTH_PIXELS;
+const LINES: i16 = crate::backend::FB_LINES as i16;
+const WIDTH: i16 = crate::backend::FB_WIDTH_PIXELS as i16;
 const WORDS_PER_LINE: usize = crate::backend::FB_WIDTH_WORDS;
-pub type FrBuf = [u32; WORDS_PER_LINE * LINES];
+pub type FrBuf = [u32; WORDS_PER_LINE * LINES as usize];
 
 // add more fonts (an example):
 // https://github.com/samblenny/blitstr2/commit/bb7d4ab6a2d8913dcb520895a3c242c933413aae
